starded developing overlapping algorithm

+++ src/main/java/pl/mdj/rejestrbiurowy/model/entity/Car.java
@@ -21,9 +21,10 @@ public class Car {
     private String registration;
     @Enumerated(EnumType.STRING)
     private CarFuel fuel;
+
     private LocalDate insuranceExpiration;
     private LocalDate inspectionExpiration;
-    private int mileage;
+    private Long mileage;
 
     public Car(String name, String registration) {
         this.name = name;
